requirements.txt
Django==4.1.4
bitrix24-rest==1.1.1
bs4==0.0.1
lxml==4.9.2
gunicorn==20.1.0
psycopg2==2.9.5
